finish general outline of changes to include system
