+++ implementation/objects_and_calls/evaluator_basics/type_operation_info.h
@@ -200,189 +200,6 @@ namespace expr
 
 		};
 
-		template <typename t>
-		struct type_operation_info<std::unique_ptr<t>> : public wrapper_type_operation_trait
-		{
-
-			template<typename name_generator = compiler_name_generator>
-			static std::string type_name(name_generator instance)
-			{
-				return instance.template retrieve<t>() + "-unique_ptr";
-			}
-
-			static std::optional<std::unique_ptr<t>> parse(std::string::const_iterator& start, std::string::const_iterator stop)
-			{
-				if (start == stop)
-				{
-					//invalid input
-					return std::optional<std::unique_ptr<t>>(nullptr);
-				}
-
-				static char const none[] = "nullptr";
-
-				auto m = std::mismatch(start, stop, std::begin(none), std::end(none) - 1);
-
-				if (m.second == std::end(none) - 1)
-				{
-					start = m.first;
-					return std::optional<std::unique_ptr<t>>(std::unique_ptr<t>(nullptr));
-				}
-				else if (*start == '&')
-				{
-					++start;
-					std::optional<t>&& g = type_operation_info<t>::parse(start, stop);
-					if (g)
-					{
-						return std::optional<std::unique_ptr<t>>(std::make_unique<t>(std::move(*g)));
-					}
-					else
-					{
-						return std::nullopt;
-					}
-				}
-				else
-				{
-					return std::nullopt;
-				}
-			}
-
-			static std::string print(std::unique_ptr<t> const& tar)
-			{
-				if (tar)
-				{
-					return "&" + type_operation_info<t>::print(*tar);
-				}
-				else
-				{
-					return "nullptr";
-				}
-			}
-
-		};
-
-		template <typename t>
-		struct type_operation_info<std::shared_ptr<t>> : public wrapper_type_operation_trait
-		{
-
-			template<typename name_generator = compiler_name_generator>
-			static std::string type_name(name_generator instance)
-			{
-				return instance.template retrieve<t>() + "-shared_ptr";
-			}
-
-			static std::optional<std::shared_ptr<t>> parse(std::string::const_iterator& start, std::string::const_iterator stop)
-			{
-				if (start == stop)
-				{
-					//invalid input
-					return std::optional<std::shared_ptr<t>>(nullptr);
-				}
-
-				static char const none[] = "nullptr";
-
-				auto m = std::mismatch(start, stop, std::begin(none), std::end(none) - 1);
-
-				if (m.second == std::end(none) - 1)
-				{
-					start = m.first;
-					return std::optional<std::shared_ptr<t>>(std::shared_ptr<t>(nullptr));
-				}
-				else if (*start == '&')
-				{
-					++start;
-					std::optional<t>&& g = type_operation_info<t>::parse(start, stop);
-					if (g)
-					{
-						return std::optional<std::shared_ptr<t>>(std::make_unique<t>(std::move(*g)));
-					}
-					else
-					{
-						return std::nullopt;
-					}
-				}
-				else
-				{
-					return std::nullopt;
-				}
-			}
-
-			static std::string print(std::shared_ptr<t> const& tar)
-			{
-				if (tar)
-				{
-					return "&" + type_operation_info<t>::print(*tar);
-				}
-				else
-				{
-					return "nullptr";
-				}
-			}
-		};
-
-
-		template <typename t>
-		struct type_operation_info<std::optional<t>> : public wrapper_type_operation_trait
-		{
-
-			static std::string print(std::optional<t> const& tar)
-			{
-				if (tar)
-				{
-					return "?" + type_operation_info<t const&>::print(*tar);
-				}
-				else
-				{
-					return "nullopt";
-				}
-			}
-
-
-
-			template<typename name_generator = compiler_name_generator>
-			static std::string type_name(name_generator instance)
-			{
-				return instance.template retrieve<t>() + "-optional";
-			}
-
-
-
-			static std::optional<std::optional<t>> parse(std::string::const_iterator& start, std::string::const_iterator stop)
-			{
-				if (start == stop)
-				{
-					//invalid input
-					return std::optional<std::optional<t>>(std::nullopt);
-				}
-
-				static char const none[] = "nullopt";
-
-				auto m = std::mismatch(start, stop, std::begin(none), std::end(none) - 1);
-
-				if (m.second == std::end(none) - 1)
-				{
-					//valid input indicating an empty optional of t.
-					start = m.first;
-					return std::optional<std::optional<t>>(std::optional<t>(std::nullopt));
-				}
-				else if (*start == '?')
-				{
-					++start;
-					std::optional<t>&& g = type_operation_info<t>::parse(start, stop);
-					if (g)
-					{
-						return std::optional<std::optional<t>>(std::move(g));
-					}
-					else
-					{
-						return std::nullopt;
-					}
-				}
-				else
-				{
-					return std::nullopt;
-				}
-			}
-		};
 
 		template<typename t>
 		class strong
@@ -455,139 +272,6 @@ namespace expr
 		};
 
 
-		template<typename t>
-		struct type_operation_info<std::vector<t> const&>
-		{
-			template<typename name_generator = compiler_name_generator>
-			static std::string type_name(name_generator instance)
-			{
-				return instance.template retrieve<t>() + "-vector";
-			}
-
-			static std::optional<std::vector<t>> parse(std::string::const_iterator& start, std::string::const_iterator stop)
-			{
-
-				if (start == stop)
-				{
-					return { std::vector<t>{} };
-				}
-				while (start != stop && *start == ' ')
-				{
-					++start;
-				}
-				if (start == stop)
-				{
-					return { std::vector<t>{} };
-				}
-				if (*start == '[' || *start == '(' || *start == '{')
-				{
-					char close_char = '!'; //not initialized
-					if (*start == '[')
-					{
-						close_char = ']';
-					}
-					else if (*start == '(')
-					{
-						close_char = ')';
-					}
-					else if (*start == '{')
-					{
-						close_char = '}';
-					}
-					std::vector<t> ret;
-					++start;
-					while (start != stop)
-					{
-						while (start != stop && *start == ' ')
-						{
-							++start;
-						}
-
-						if (start == stop)
-						{
-							return std::nullopt;
-						}
-						else if (*start == ']')
-						{
-							++start;
-							return { std::move(ret) };
-						}
-
-						std::optional<t> curelem = type_operation_info<t>::parse(start, stop);
-						if (curelem)
-						{
-							ret.emplace_back(std::move(*curelem));
-						}
-						else
-						{
-							return std::nullopt;
-						}
-
-						while (start != stop && *start == ' ')
-						{
-							++start;
-						}
-						if (start == stop)
-						{
-							return std::nullopt;
-						}
-						if (*start == close_char)
-						{
-							++start;
-							return { std::move(ret) };
-						}
-						continue; //the list does not require commas.
-					}
-					return std::nullopt;
-				}
-				else
-				{
-					return std::nullopt;
-				}
-
-			}
-
-			static std::string print(std::vector<t> const& tar)
-			{
-				std::string ret;
-				ret.push_back('[');
-				for (auto& it : tar)
-				{
-					ret.append(type_operation_info<t>::print(it));
-					ret.push_back(' ');
-				}
-				if (ret.size() == 1)
-				{
-					ret.push_back(' ');
-				}
-				(*ret.rbegin()) = ']';
-				return std::move(ret);
-			}
-		};
-
-		template<>
-		struct type_operation_info<std::string const&>
-		{
-			template<typename name_generator = compiler_name_generator>
-			static std::string type_name(name_generator instance)
-			{
-				return "string";
-			}
-
-			static std::optional<std::string> parse(std::string::const_iterator& start, std::string::const_iterator stop)
-			{
-				return std::optional<std::string>{ std::string(start,stop) };
-			}
-
-			static std::string print(std::string const& a)
-			{
-				return std::string(a);
-			}
-		};
-
-
-
-
 
 		template<typename t, typename enable = void>
 		struct wrapper_type_operation_info
@@ -598,33 +282,6 @@ namespace expr
 			}
 		};
 
-		template<typename t>
-		struct wrapper_type_operation_info<t,std::enable_if_t<std::is_base_of_v<wrapper_type_operation_trait, type_operation_info<t>>>>
-		{
-			static constexpr bool is_wrapper()
-			{
-				return true;
-			}
-
-			using wrapped = decltype(*std::declval<t>());
-
-			static bool can_unwrap(t&& a)
-			{
-				if (std::forward<t>(a))
-				{
-					return true;
-				}
-				else
-				{
-					return false;
-				}
-			}
-
-			static wrapped do_unwrap(t&& a)
-			{
-				return *std::forward<t>(a);
-			}
-		};
