+++ include/features/all_traits.h
@@ -1,0 +1,1 @@
+#include"wrapper_trait.h"
