--- include/quick_evaluator.h
@@ -1,23 +1,0 @@
-#pragma once
-
-#include"evaluator_basics.h"
-
-namespace expr
-{
-
-	//core utilities to the evaluator
-	using impl::core;
-
-	//core utilities specific to c++
-	using impl::cpp_core;
-
-	//utilities for interacting with the terminal and system
-	using impl::terminal;
-
-	//utilities generic to any type
-	using impl::util;
-
-	//utilities for wrappers of any type
-	using impl::wrapper_util;
-}
-
